// src/types/coordinates.ts
export type CoordinateLLH = {
    lat: number;
    lon: number;
    h: number;
  };
  
  export type CoordianteXYZ = {
    x: number;
    y: number;
    z: number;
  };
  
  export type CoordinateENU = {
    E: number;
    N: number;
    U: number;
  };
  
  export type CoordinateAER = {
    az: number;
    elev: number;
    range: number;
  };

  export type Hms = {
    h: number;
    m: number;
    s: number;
  };
  